+++ docs/source/_static/export_all_data.py
@@ -1,20 +1,27 @@
 #!/usr/bin/env python
 
-"""export_all_data.py - script for exporting all available data"""
+"""
+export_all_data.py - script for exporting available data.
+"""
 
 import os
+import argparse
+
 from collectionbatchtool import *
 
 
-def export_all_data(output_dir=None, quiet=True):
+def export_all_data(
+        config_file, output_dir=None,
+        drop_empty_columns=False, quiet=True):
     """
-    Export table data to CSV files.
+    Export data from Specify to CSV files.
 
     Parameters
     ----------
     output_dir : str
         Path to the output directory.
     """
+    apply_user_settings(config_file) 
     output_dir = output_dir if output_dir else ''
     for tabledataset_subclass in TableDataset.__subclasses__():
         instance = tabledataset_subclass()
@@ -23,9 +30,36 @@ def export_all_data(output_dir=None, quiet=True):
             filename = instance.model.__name__.lower() + '.csv'
             filepath = os.path.join(output_dir, filename)
             instance.to_csv(
-                filepath, update_sourceid=True, quiet=quiet)
+                filepath,
+                update_sourceid=True,
+                drop_empty_columns=drop_empty_columns,
+                quiet=quiet)
 
 
 if __name__ == '__main__':
-    apply_user_settings('settings.cfg')  # change to your own config-file!
-    export_all_data(quiet=False)  # call the export function
+    parser = argparse.ArgumentParser(
+        description='A command-line script to export Specify data.')
+    parser.add_argument(
+        'config_file', type=argparse.FileType('rU'),
+        help='path to a config-file')
+    parser.add_argument(
+        'output_dir', default='.', nargs='?',
+        help='path to output directory')
+    parser.add_argument(
+        '-d', '--drop-empty-columns', dest='drop_empty_columns',
+        action='store_true', help='drop columns without data')
+    parser.add_argument(
+        '-v', '--verbose', action='store_true')
+    args = parser.parse_args()
+ 
+    if not os.path.isdir(args.output_dir):
+        msg = "%d is not valid directory" % args.output_dir
+        raise argparse.ArgumentTypeError(msg)
+    
+    quiet = False if args.verbose else True
+ 
+    export_all_data(
+        args.config_file.name,
+        output_dir=args.output_dir,
+        drop_empty_columns=args.drop_empty_columns,
+        quiet=quiet)